Return FPS as a float value for a little more accuracy.

+++ src/game.h
@@ -32,7 +32,7 @@ void gshmup_init_game (void);
 void gshmup_run_game (void);
 void gshmup_set_current_scene (GshmupScene *scene);
 void gshmup_set_current_agenda (SCM agenda);
-int gshmup_get_fps (void);
+float  gshmup_get_fps (void);
 void gshmup_game_init_scm (void);
 bool gshmup_debug_mode (void);
 SCM gshmup_create_agenda (void);
